Pull `init` `Run` method into new method in a separate file, in preparation for adding experimental fork.
